# Server/openapi.yaml
openapi: 3.0.0
info:
  title: TestVar - Flashcards API
  description: "A revolutionary REST API for flashcards"
  version: 1.0.0
servers:
  - url: http://localhost:7222/api
    description: Development server
paths:
...
  /users:
    get:
      summary: "Get all users"
      tags:
        - "Users"
      responses:
        "200":
          description: "A list of users"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/User"
    post:
      summary: "Create a new user"
      tags:
        - "Users"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/User"
                - type: object
                  properties:
                    password:
                      type: "string"
                      format: "password"
                      example: "password"
                      nullable: true
      responses:
        "201":
          description: "The created user"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: "The user could not be created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        description: "The ID of the user"
        schema:
          type: "string"
    get:
      summary: "Get a user by ID"
      tags:
        - "Users"
      responses:
        "200":
          description: "The user"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: "The user was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    put:
      summary: "Update a user by ID"
      tags:
        - "Users"
      description: "Admin status only editable if the current user is an admin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/User"
                - type: object
                  properties:
                    password:
                      type: "string"
                      format: "password"
                      example: "password"
                      nullable: true
      responses:
        "200":
          description: "The updated user"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: "The user was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    delete:
      summary: "Delete a user by ID"
      tags:
        - "Users"
      responses:
        "204":
          description: "The user was deleted"
        "404":
          description: "The user was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /users/{userID}/sets:
    get:
      summary: "Get all flashcard sets created by a user"
      tags:
        - "Users"
        - "FlashCard Sets"
      parameters:
        - name: userID
          in: path
          required: true
          description: "The ID of the user"
          schema:
            type: "string"
      responses:
        "200":
          description: "A list of flashcard sets"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/FlashcardSet"
        "404":
          description: "The user was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /users/{userID}/collections/{collectionID}:
    parameters:
      - name: userID
        in: path
        required: true
        description: "The ID of the user"
        schema:
          type: "string"
      - name: collectionID
        in: path
        required: true
        description: "The ID of the collection"
        schema:
          type: "string"
    get:
      summary: "Get a flashcard set collection by ID"
      tags:
        - "Collections"
      responses:
        "200":
          description: "The flashcard set collection"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "404":
          description: "The flashcard set collection was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    put:
      summary: "Update a flashcard set collection by ID"
      tags:
        - "Collections"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                comment:
                  type: "string"
                  nullable: false
                  example: "I love this set!"
                sets:
                  type: "array"
                  nullable: true
                  items:
                    type: "number"
                    description: "The IDs of member flashcards"
                    example: 1
      
      responses:
        "200":
          description: "The updated flashcard set collection"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "404":
          description: "The flashcard set collection was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: "You are not authorized ('logged in') to update this collection"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: "You are not allowed to update this collection"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    delete:
      summary: "Delete a flashcard set collection by ID"
      tags:
        - "Collections"
      responses:
        "204":
          description: "The flashcard set collection was deleted"
        "404":
          description: "The flashcard set collection was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: "You are not authorized ('logged in') to delete this collection"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        '403':
          description: "You are not allowed to delete this collection"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
components:
  schemas:
...
    FlashCard:
      type: "object"
      properties:
        question:
          type: "string"
          example: "What is the capital of France?"
        answer:
          type: "string"
          example: "Paris"
        difficulty:
          $ref: "#/components/schemas/Difficulty"
          nullable: true
    FlashcardSet:
      type: "object"
      properties:
        id:
          type: "number"
          example: 1
        name:
          type: "string"
          example: "European Capitals"
        cards:
          type: "array"
          items:
            $ref: "#/components/schemas/FlashCard"
        created_at:
          type: "string"
          format: "date-time"
          example: "2024-10-14T12:00:00Z"
        updated_at:
          type: "string"
          format: "date-time"
          example: "2024-10-15T12:00:00Z"
    ProblemDetails:
      type: "object"
      properties:
        title:
          type: "string"
          nullable: false
          description: "Summary of the problem"
        status:
          type: "number"
          nullable: false
          description: "HTTP status code"
        detail:
          type: "string"
          nullable: true
          description: "Additional problem details"
    Difficulty:
      type: "string"
      enum:
        - "easy"
        - "medium"
        - "hard"
    User:
      type: "object"
      properties:
        id:
          type: "number"
          example: 1
        username:
          type: "string"
          example: "johndoe"
        admin:
          type: "boolean"
          example: false
          default: false
          nullable: true
          description: "The set creation limit should only apply to non-admin users"
...
    Collection:
      type: "object"
      properties:
        id:
          type: "number"
          example: 1
        comment:
          type: "string"
          example: "I like these sets"
        owner:
          $ref: "#/components/schemas/User"
        sets:
          type: "array"
          items:
            $ref: "#/components/schemas/FlashcardSet"